% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{annot2dpv}
\alias{annot2dpv}
\title{Convert annotation file to dpv}
\usage{
annot2dpv(
  input_file,
  output_file,
  coordinates = NULL,
  indeces = NULL,
  verbose = TRUE
)
}
\arguments{
\item{input_file}{annotation file path}

\item{output_file}{path to output file}

\item{coordinates}{contains the vertex coordinates or face indices}

\item{indeces}{vertex or face sequential index}

\item{verbose}{logical indicating to be verbose or not}
}
\value{
data frame
}
\description{
Based on matlab scripts from
Anderson M. Winkler
Yale University / Institute of Living
Oct/2011
http://brainder.org
}
