% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_annotation.R
\name{annotation}
\alias{annotation}
\title{Constructor for freesurfer_annotation-class}
\usage{
annotation(vertices, labels, colortable = NULL)
}
\arguments{
\item{vertices}{vector of vertices}

\item{labels}{vector of labels}

\item{colortable}{color table}
}
\description{
freesurfer_annotation is a special object
containing the three components of a 
FreeSurfer annotation file: vertices, labels
and the colortable
}
