% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-ggseg3d-atlas.R
\name{aparc_2_mesh}
\alias{aparc_2_mesh}
\title{Converts annotations fo atlas}
\usage{
aparc_2_mesh(
  subject = "fsaverage5",
  hemisphere = "rh",
  surface = "inflated",
  annot = "aparc",
  subjects_dir = freesurfer::fs_subj_dir(),
  annot_dir = file.path(subjects_dir, subject, "label"),
  output_dir = tempdir(),
  cleanup = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{subject}{Freesurfer subject, must exist in whatever subject directory
specified or set in the environment with $SUBJECTS_DIR}

\item{hemisphere}{hemisphere, either "rh" or "lh"}

\item{surface}{surface from subjects surf-folder}

\item{annot}{base-name of annot file}

\item{subjects_dir}{Freesurfer subject directory}

\item{annot_dir}{path to directory with annot-files}

\item{output_dir}{output directory path}

\item{cleanup}{logical to toggle removal of all intermediary files}

\item{verbose}{logical indicating to be verbose or not}
}
\value{
data frame with one row per label
}
\description{
This function will create an atlas
ready data.frame for ggseg3d to
plot with plotly.
}
\details{
Based on A. Winkler scripts
}
\examples{
\dontrun{
dt <- aparc_2_mesh()
dt <- aparc_2_mesh(surface = "white")
dt <- aparc_2_mesh(hemisphere = "lh")
dt <- aparc_2_mesh(annot = "aparc.a2009s")
}

}
