% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{asc2ply}
\alias{asc2ply}
\title{Convert ascii to .ply}
\usage{
asc2ply(input_file, output_file = gsub("\\\\.dpv", ".ply", input_file))
}
\arguments{
\item{input_file}{input surface or curvature file made with 
\code{\link{surf2asc}}, \code{\link{curv2asc}}, or \code{\link{curvnf2asc}}}

\item{output_file}{path to output file}
}
\value{
ply text file
}
\description{
ggseg3d bases its functions on data
from .ply files. To turn Freesurfer 
surface or curvature based ascii files
into ply, this is the function you need.
}
