% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-steps.R
\name{atlas_isolate}
\alias{atlas_isolate}
\title{Isolate raster region}
\usage{
atlas_isolate(outdir, dilation = 2, eroding = 2, smoothing = 4, verbose = TRUE)
}
\arguments{
\item{outdir}{output directory}

\item{dilation}{raster dilation}

\item{eroding}{raster eroding}

\item{smoothing}{raster smoothing}

\item{verbose}{be verbose}
}
\description{
Isolate raster region
}
