% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-steps.R
\name{atlas_lab2ctab}
\alias{atlas_lab2ctab}
\title{Label to ctab}
\usage{
atlas_lab2ctab(outdir, verbose)
}
\arguments{
\item{outdir}{output directory}

\item{verbose}{be verbose}
}
\description{
Label to ctab
}
