% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-steps.R
\name{atlas_labelgii}
\alias{atlas_labelgii}
\title{label to gifti}
\usage{
atlas_labelgii(outdir, annotdir)
}
\arguments{
\item{outdir}{output directory}

\item{annotdir}{annotation directory}
}
\description{
label to gifti
}
