% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-steps.R
\name{atlas_raster}
\alias{atlas_raster}
\title{Rasterise images}
\usage{
atlas_raster(indir, verbose = TRUE)
}
\arguments{
\item{indir}{input directory}

\item{verbose}{be verbose}
}
\description{
Rasterise images
}
