% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-steps.R
\name{atlas_raster2sf}
\alias{atlas_raster2sf}
\title{Raster to sf}
\usage{
atlas_raster2sf(rasterobjs, verbose = TRUE)
}
\arguments{
\item{rasterobjs}{raster objects}

\item{verbose}{be verbose}
}
\description{
Raster to sf
}
