% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-steps.R
\name{atlas_sf2gg}
\alias{atlas_sf2gg}
\title{sf to ggplot}
\usage{
atlas_sf2gg(atlas_df, atlas_name, verbose = TRUE)
}
\arguments{
\item{atlas_df}{sf data frame}

\item{atlas_name}{name of atlas}

\item{verbose}{be verbose}
}
\description{
sf to ggplot
}
