% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-steps.R
\name{atlas_tcl}
\alias{atlas_tcl}
\title{Run tcl script}
\usage{
atlas_tcl(annot_lab, outdir, verbose)
}
\arguments{
\item{annot_lab}{annotation label}

\item{outdir}{output directory}

\item{verbose}{be verbose}
}
\description{
Run tcl script
}
