% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-steps.R
\name{atlas_vol2label}
\alias{atlas_vol2label}
\title{Volume to label}
\usage{
atlas_vol2label(annot_lab, outdir, verbose)
}
\arguments{
\item{annot_lab}{annotation label}

\item{outdir}{output directory}

\item{verbose}{be verbose}
}
\description{
Volume to label
}
