% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-steps.R
\name{atlas_vol2surf}
\alias{atlas_vol2surf}
\title{Nifti volume to surface}
\usage{
atlas_vol2surf(infile, outdir, projfrac = 0.5, verbose)
}
\arguments{
\item{infile}{nifti volume}

\item{outdir}{output directory}

\item{projfrac}{value to mri_vol2surf -projfrac}

\item{verbose}{be verbose}
}
\description{
Nifti volume to surface
}
