% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_pals.R
\docType{data}
\name{brain_pals}
\alias{brain_pals}
\title{ggseg palettes}
\format{A list of palettes for each atlas in ggseg.}
\usage{
data(brain_pals)
}
\description{
Appends palettes from the ggsegExtra package to the
original palettes from the ggseg-package
}
\keyword{palettes}
