% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract-geometry.R
\name{center_meshes}
\alias{center_meshes}
\title{Center meshes around origin}
\usage{
center_meshes(meshes_list)
}
\arguments{
\item{meshes_list}{Named list of meshes}
}
\value{
Centered meshes list
}
\description{
Computes the global centroid of all mesh vertices and translates
all meshes to center around the origin.
}
\keyword{internal}
