% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{check_fs}
\alias{check_fs}
\title{Check if FS can be run}
\usage{
check_fs(msg = NULL, abort = FALSE)
}
\arguments{
\item{msg}{message to print on error}

\item{abort}{logical. If function should error
if Freesurfer is not installed. Defaults to FALSE.}
}
\value{
logical
}
\description{
Check if FS can be run
}
\keyword{internal}
