% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chen.R
\docType{data}
\name{chenTh}
\alias{chenTh}
\title{Genetic topography of brain thickness morphology}
\format{A data.frame with 11341 observations and 11 variables
\describe{
  \item{long}{coordinates for the x-axis}
  \item{lat}{coordinates for the y-axis}
  \item{area}{name of region}
  \item{hemi}{dummy name of the hemisphere}
  \item{side}{which side to view (sagittal)}
  \item{label}{label name from Freesurfer}
  \item{atlas}{name of the atlas}
}}
\usage{
chenTh
}
\description{
Genetic topography of brain thickness morphology
}
\examples{
data(chenTh)
}
\references{
Chen et al. (2013) PNAS, 110 (42) 17089-17094; 
(\href{https://doi.org/10.1073/pnas.1308091110 }{PubMed})
}
\keyword{datasets}
