% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-convert.R
\name{compute_vertex_indices}
\alias{compute_vertex_indices}
\title{Compute vertex indices by matching to brain mesh}
\usage{
compute_vertex_indices(
  atlas_3d_flat,
  surface = "inflated",
  subject = "fsaverage5",
  tolerance = 1e-04
)
}
\description{
Compute vertex indices by matching to brain mesh
}
\keyword{internal}
