% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colortable-funcs.R
\name{ctab_line}
\alias{ctab_line}
\title{Make a string for ctab writing}
\usage{
ctab_line(idx, name, R, G, B, A)
}
\arguments{
\item{idx}{roi index}

\item{name}{roi name}

\item{R}{roi red value}

\item{G}{roi green value}

\item{B}{roi blue value}

\item{A}{roi opacity/alpha value}
}
\description{
Make a string for ctab writing
}
