% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{curv2asc}
\alias{curv2asc}
\title{Convert Freesurfer curvature file to ascii}
\usage{
curv2asc(input_file, white, output_file, verbose = TRUE)
}
\arguments{
\item{input_file}{path to curvature file}

\item{white}{path to subjects hemi.white file}

\item{output_file}{path to output file}

\item{verbose}{logical indicating to be verbose or not}
}
\value{
ascii data
}
\description{
Curvature files from Freesurfer need to be
conserted to be handled in R. This first
step turns the curvature file in to a text
ascii file.
}
