% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{curvnf2asc}
\alias{curvnf2asc}
\title{Convert Freesurfer no fix curvature file to ascii}
\usage{
curvnf2asc(input_file, nofix, output_file, verbose = TRUE)
}
\arguments{
\item{input_file}{path to nofix curvature file}

\item{nofix}{path to subjects hemi.orig.nofix file}

\item{output_file}{path to output file}

\item{verbose}{logical indicating to be verbose or not}
}
\value{
ascii data
}
\description{
No fix curvature files from Freesurfer need to be
conserted to be handled in R. This first
step turns the no fix curvature file in to a text
ascii file.
}
