% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{detect_hemi_from_gifti_filename}
\alias{detect_hemi_from_gifti_filename}
\title{Detect hemisphere from GIFTI filename}
\usage{
detect_hemi_from_gifti_filename(filename)
}
\arguments{
\item{filename}{Basename of the GIFTI file}
}
\value{
"lh" or "rh", or NA if undetectable
}
\description{
Detect hemisphere from GIFTI filename
}
\keyword{internal}
