% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkextra.R
\docType{data}
\name{dkextra}
\alias{dkextra}
\title{Desikan-Killiany Cortical Atlas}
\format{A data.frame with 10913 observations and 11 variables
\describe{
  \item{long}{coordinates for the x-axis}
  \item{lat}{coordinates for the y-axis}
  \item{region}{name of region}
  \item{hemi}{name of the hemisphere (left, right)}
  \item{side}{which side to view (medial, lateral)}
  \item{acronym}{abbreviated name of region}
  \item{lobe}{lobe location}
  \item{label}{label from recon-all segmentation}
  \item{atlas}{name of the atlas}
}}
\usage{
data(dkextra)
}
\description{
Coordinate data for the Desikan-Killany Cortical atlas,
with 40 regions in on the cortical surface of the brain.
}
\examples{
data(dkextra)
}
\references{
Fischl et al. (2004) Cerebral Cortex 14:11-22
(\href{https://academic.oup.com/cercor/article/14/1/11/433466}{PubMed})
}
\seealso{
Other ggseg_atlases: 
\code{\link{chenAr}},
\code{\link{chenTh}},
\code{\link{glasser}},
\code{\link{hcpa_3d}},
\code{\link{hoCort}},
\code{\link{jhu}},
\code{\link{tracula}},
\code{\link{yeo17}},
\code{\link{yeo7}}
}
\concept{ggseg_atlases}
\keyword{datasets}
