% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify-atlas.R
\name{flatten_ggseg3d_atlas}
\alias{flatten_ggseg3d_atlas}
\title{Flatten ggseg3d_atlas to data frame}
\usage{
flatten_ggseg3d_atlas(atlas_3d, surface = "inflated")
}
\description{
Flatten ggseg3d_atlas to data frame
}
\keyword{internal}
