% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-structs.R
\name{fs_nofixcurv}
\alias{fs_nofixcurv}
\title{Available Freesurfer no fix curvatures}
\usage{
fs_nofixcurv()
}
\value{
character
}
\description{
Available Freesurfer no fix curvatures
}
\examples{
fs_nofixcurv()
}
