% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freesurfer.R
\name{fs_nofixcurvatures}
\alias{fs_nofixcurvatures}
\title{Available Freesurfer no fix curvatures}
\usage{
fs_nofixcurvatures()
}
\value{
character vector of no-fix curvature file names
}
\description{
Available Freesurfer no fix curvatures
}
\keyword{internal}
