% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl-utils.R
\name{get_fsl_colour}
\alias{get_fsl_colour}
\title{Extract colours for FSL atlas}
\usage{
get_fsl_colour(atlas = "HarvardOxford-Cortical", path = NULL)
}
\arguments{
\item{atlas}{string name of atlas}
}
\value{
data.frame
}
\description{
Extract colours for FSL atlas
}
