% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mesh}
\alias{get_mesh}
\title{Extract mesh data from ply}
\usage{
get_mesh(ply, ...)
}
\arguments{
\item{ply}{path to ply-file}

\item{...}{arguments to \code{\link[geomorph]{read.ply}}}
}
\value{
list of meshes and faces
}
\description{
.ply files contain a lot of information.
for ggseg3d, we only need information
on the vertices and faces of the mesh.
Thes function opens a ply file, and
organises the meshes and faces into 
a single list.
}
\examples{
\dontrun{
get_mesh("path/to/surface.ply")

# Turn off showing the ply when reading
get_mesh("path/to/surface.ply", ShowSpecimen = FALSE)
}
}
