% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_verbose}
\alias{get_verbose}
\title{Get verbose setting}
\usage{
get_verbose()
}
\value{
Logical TRUE/FALSE
}
\description{
Returns the verbose setting from option, environment variable, or default.
Checks in order: \code{ggseg.extra.verbose} option, \code{GGSEG_EXTRA_VERBOSE} env var,
then defaults to TRUE.
}
\details{
Used as default argument in verbose parameters throughout the package.
Set \code{options(ggseg.extra.verbose = FALSE)} or
\code{Sys.setenv(GGSEG_EXTRA_VERBOSE = "false")} to suppress output globally.
}
\examples{
get_verbose()
}
