% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-ggseg-atlas.R
\name{ggseg3d_2_ggseg}
\alias{ggseg3d_2_ggseg}
\title{Turn ggseg3d-atlas to ggseg}
\usage{
ggseg3d_2_ggseg(
  ggseg3d_atlas = ggseg3d::dk_3d,
  steps = 1:5,
  output_dir = tempdir(),
  cleanup = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{ggseg3d_atlas}{object of class ggseg3d-atlas}

\item{steps}{number of 1:5 of which steps to run through}

\item{output_dir}{output directory path}

\item{cleanup}{logical to toggle removal of all intermediary files}

\item{verbose}{logical indicating to be verbose or not}
}
\value{
data.frame ready for manual cleaning
}
\description{
Function will create a dataframe
based on a ggseg3d atlas, based on 
the contours of each segment.
}
