% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegExtra-package.R
\docType{package}
\name{ggsegExtra-package}
\alias{ggsegExtra}
\alias{ggsegExtra-package}
\title{ggsegExtra: Extra Utilities for the 'ggseg' and 'ggseg3d' Plotting Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Contains functions to create new data sets compatible with the 'ggseg' and 'ggseg3d' packages for plotting brain segmentations through R. Requires several external software to be installed on the system for atlas creations to work, and most of these are not available for windows. But atlas creation tools exist for 2d simple features data, and also 3d tri-surface mesh plots. Also has a list of compatible data sets and functions for easy installation of these.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggseg/ggsegExtra}
  \item Report bugs at \url{https://github.com/ggseg/ggsegExtra/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Didac Vidal-Piñeiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

Other contributors:
\itemize{
  \item John Muschelli \email{muschellij2@gmail.com} (\href{https://orcid.org/0000-0001-6469-1750}{ORCID}) [contributor]
}

}
\keyword{internal}
