% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_pa.R
\docType{data}
\name{hc_pa_3d}
\alias{hc_pa_3d}
\title{Freesurfer ASEG with posterior-anterior hippocampus}
\format{A data.frame with 2702 observations and 9 variables
\describe{
  \item{lat}{coordinates for the x-axis}
  \item{long}{coordinates for the y-axis}
  \item{area}{acronym of network}
  \item{name}{full name of network}
  \item{hemi}{name of the hemisphere (left, right)}
  \item{side}{which side to view (medial, lateral,axial)}
  \item{atlas}{name of the atlas}
}}
\usage{
data(hc_pa_3d)
}
\description{
Coordinate data for the subcortical parcellations implemented
in Freesurfer, with a division of the hippocampus in posterior
and anterior segments.
}
\examples{
data(hc_pa_3d)
}
\references{
Fischl et al., (2002). Neuron, 33:341-355
(\href{https://www.ncbi.nlm.nih.gov/pubmed/11832223}{PubMed})
}
\seealso{
Other ggseg_atlases: 
\code{\link{chenAr}},
\code{\link{chenTh}},
\code{\link{glasser}},
\code{\link{hoCort}},
\code{\link{jhu}},
\code{\link{tracula}},
\code{\link{yeo17}},
\code{\link{yeo7}}
}
\concept{ggseg_atlases}
\keyword{datasets}
