% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icbm.R
\docType{data}
\name{icbm_3d}
\alias{icbm_3d}
\title{ICBM white matter parcellation}
\format{An object of class \code{ggseg_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.}
\usage{
icbm_3d
}
\description{
ICBM white matter parcellation
}
\examples{
data(icbm_3d)
}
\references{
Mori et al., MRI Atlas of Human White Matter. Elsevier, Amsterdam, The Netherlands (2005) 
(\href{https://www.elsevier.com/books/mri-atlas-of-human-white-matter/mori/978-0-444-51741-8}{PubMed})
}
\seealso{
Other ggseg3d_atlases: \code{\link{desterieux_3d}},
  \code{\link{glasser_3d}}, \code{\link{hc_pa_3d}},
  \code{\link{jhu_3d}}, \code{\link{schaefer17_3d}},
  \code{\link{schaefer7_3d}}, \code{\link{tracula_3d}},
  \code{\link{yeo17_3d}}, \code{\link{yeo7_3d}}
}
\concept{ggseg3d_atlases}
\keyword{datasets}
