% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-atlases.R
\name{install_ggseg_atlas}
\alias{install_ggseg_atlas}
\title{Install a ggseg atlas package}
\usage{
install_ggseg_atlas(
  package,
  repos = c(ggseg = "https://ggseg.r-universe.dev", getOption("repos")),
  ...
)
}
\arguments{
\item{package}{Package name (e.g., \code{"ggsegYeo2011"}, \code{"ggsegDKT"}).
Use \code{\link[=ggseg_atlas_repos]{ggseg_atlas_repos()}} to see available packages.}

\item{repos}{Repositories to install from. Defaults to the ggseg
r-universe plus your current CRAN mirror.}

\item{...}{Additional arguments passed to \code{\link[utils:install.packages]{utils::install.packages()}}.}
}
\description{
Install an atlas package from the ggseg r-universe. This is the
easiest way to get pre-built atlases like DKT, Yeo, Schaefer, etc.
}
\examples{
\dontrun{
# Find available Yeo atlases
ggseg_atlas_repos("yeo")

# Install one
install_ggseg_atlas("ggsegYeo2011")
}
}
\seealso{
\code{\link[=ggseg_atlas_repos]{ggseg_atlas_repos()}} to list available atlases
}
