% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{is_ctab}
\alias{is_ctab}
\title{Check if object is a color table}
\usage{
is_ctab(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
TRUE if x is a data.frame with the required color table columns.
}
\description{
Check if object is a color table
}
