% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_facewise}
\alias{is_facewise}
\title{Check if dpv is facewise}
\usage{
is_facewise(dpx, surface, verbose = TRUE)
}
\arguments{
\item{dpx}{vertex or face data}

\item{surface}{surface vertex and face list}

\item{verbose}{logical indicating to be verbose or not}
}
\description{
Check if dpv is facewise
}
