% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{is_volume_file}
\alias{is_volume_file}
\title{Check if file is a supported volume format}
\usage{
is_volume_file(file)
}
\arguments{
\item{file}{Path to file}
}
\value{
Logical
}
\description{
Check if file is a supported volume format
}
\keyword{internal}
