% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isolate_region}
\alias{isolate_region}
\title{Isolate region to alpha channel}
\usage{
isolate_region(input_file, output_file, interrim_file = tempfile())
}
\arguments{
\item{input_file}{image file path}

\item{output_file}{output file path}

\item{interrim_file}{interrim image path}
}
\description{
Isolate region to alpha channel
}
