% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu.R
\docType{data}
\name{jhu}
\alias{jhu}
\title{JHU parcellation}
\format{A data.frame with 11341 observations and 11 variables
\describe{
  \item{long}{coordinates for the x-axis}
  \item{lat}{coordinates for the y-axis}
  \item{area}{name of region}
  \item{hemi}{dummy name of the hemisphere}
  \item{side}{which side to view (sagittal)}
  \item{label}{label name from Freesurfer}
  \item{atlas}{name of the atlas}
}}
\usage{
jhu
}
\description{
JHU parcellation
}
\examples{
data(jhu)
}
\references{
Hua et al. (2008) NeuroImage, 39(1):336-347 
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2724595/}{PubMed})
}
\seealso{
Other ggseg_atlases: 
\code{\link{chenAr}},
\code{\link{chenTh}},
\code{\link{dkextra}},
\code{\link{glasser}},
\code{\link{hcpa_3d}},
\code{\link{hoCort}},
\code{\link{tracula}},
\code{\link{yeo17}},
\code{\link{yeo7}}
}
\concept{ggseg_atlases}
\keyword{datasets}
