% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu.R
\docType{data}
\name{jhu_3d}
\alias{jhu_3d}
\title{Parcellation from JHU}
\format{An object of class \code{ggseg_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.}
\usage{
jhu_3d
}
\description{
Parcellation from JHU
}
\examples{
data(jhu_3d)
}
\references{
Hua et al., Tract probability maps in stereotaxic spaces:
analysis of white matter anatomy and tract-specific quantification. NeuroImage, 39(1):336-347 (2008)
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2724595/}{PubMed})
}
\seealso{
Other ggseg3d_atlases: 
\code{\link{desterieux_3d}},
\code{\link{glasser_3d}},
\code{\link{icbm_3d}},
\code{\link{schaefer17_3d}},
\code{\link{schaefer7_3d}},
\code{\link{tracula_3d}},
\code{\link{yeo17_3d}},
\code{\link{yeo7_3d}}
}
\concept{ggseg3d_atlases}
\keyword{datasets}
