% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-ggseg3d-atlas.R
\name{make_aseg_2_3datlas}
\alias{make_aseg_2_3datlas}
\title{Subcortical segmentation to atlas}
\usage{
make_aseg_2_3datlas(
  subject = "fsaverage5",
  subjects_dir = freesurfer::fs_subj_dir(),
  template = file.path(subjects_dir, subject, "mri/aseg.mgz"),
  color_lut = file.path(freesurfer::fs_dir(), "ASegStatsLUT.txt"),
  steps = 1:5,
  output_dir = tempdir(),
  verbose = TRUE,
  cleanup = TRUE
)
}
\arguments{
\item{subject}{Freesurfer subject, must exist in whatever subject directory
specified or set in the environment with $SUBJECTS_DIR}

\item{subjects_dir}{Freesurfer subject directory}

\item{template}{tempalte volume.mgz file path}

\item{color_lut}{Freesurfer colour look-up-table. Either
as a path or a data.frame that \code{\link{is_ctab}}}

\item{steps}{if cleanup is disabled, all files are
saved, and steps can be re-run individually}

\item{output_dir}{output directory path}

\item{verbose}{logical indicating to be verbose or not}

\item{cleanup}{logical to toggle removal of all intermediary files}
}
\value{
returns a ggseg3d-atlas ready object
}
\description{
Subcortical segmentation to atlas
}
