% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-ggseg-atlas.R
\name{make_ggseg3d_2_ggseg}
\alias{make_ggseg3d_2_ggseg}
\title{Turn ggseg3d-atlas to ggseg}
\usage{
make_ggseg3d_2_ggseg(
  ggseg3d_atlas,
  steps = 1:7,
  output_dir = tempdir(),
  tolerance = 0,
  smoothness = 5,
  cleanup = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{ggseg3d_atlas}{object of class ggseg3d-atlas}

\item{steps}{numeric vector of steps to run}

\item{output_dir}{output directory path}

\item{tolerance}{tolerance during vertex reduction \code{\link[sf]{st_simplify}}}

\item{smoothness}{smoothing factor, argument to \code{\link[smoothr]{smooth}}}

\item{cleanup}{logical to toggle removal of all intermediary files}

\item{ncores}{number of cores for parallel processing (default numcores-2)}
}
\value{
data.frame ready for manual cleaning before turning into a proper ggseg3d-atlas
}
\description{
Function will create a data.frame
based on a ggseg3d atlas, based on
the contours of each segment.
}
\examples{
\dontrun{

# Create the DKT atlas as found in the FreeSurfer Subjects directory
# And output the temporary files to the Desktop.
dkt_3d <- make_aparc_2_3datlas(annot = "aparc.DKTatlas",
    output_dir = "~/Desktop/")
}
}
