% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribute-atlas.R
\name{make_ggseg_repo}
\alias{make_ggseg_repo}
\title{Initiate a ggseg atlas repo}
\usage{
make_ggseg_repo(
  atlas_name,
  package_name,
  directory = package_name,
  git_init = TRUE,
  remote_repo = NULL
)
}
\arguments{
\item{atlas_name}{name for atlases, should be lower snake-case}

\item{package_name}{package name in snake case}

\item{directory}{directory to place the atlas in, defaults to package name}

\item{git_init}{should git repo be initiated}

\item{remote_repo}{a remote git repository to push to (must already exist)}
}
\description{
This function will clone an existing
ggseg atlas, and rename files and substitute
strings in files to prepare a repository for
ggseg atlas constribution.
}
\details{
This process is not perfect, and you will need
to go through files and make sure the information
is correct, add data to the data-folder, alter
atlas description in R-folder, make sure
tests run in tests/testthat (comment out what is
not necessary), etc. 

At the end, a complete atlas ready for addition to
ggsegExtra has noe errors and no warnings,
but notes on data size are acceptable.
}
