% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-ggseg-atlas.R
\name{make_palette_ggseg}
\alias{make_palette_ggseg}
\title{Create ggseg palette from ggseg3d-atlas}
\usage{
make_palette_ggseg(ggseg3d_atlas)
}
\arguments{
\item{ggseg3d_atlas}{object of class ggseg3d-atlas}
}
\value{
list
}
\description{
atlases in ggseg have palettes based on
colours from the paper originally
introducing the atlas. These
colours are hard-coded into ggseg3d-atlases.
This function extracts those and makes a object
ready for incporporation to a ggseg-atlas repository
}
\examples{
make_palette_ggseg(ggseg3d::dk_3d)
}
