% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-ggseg3d-atlas.R
\name{make_volumetric_2_3datlas}
\alias{make_volumetric_2_3datlas}
\title{Volumetric segmentation to 3d-atlas}
\usage{
make_volumetric_2_3datlas(
  subject = "fsaverage5",
  subjects_dir = freesurfer::fs_subj_dir(),
  template = file.path(subjects_dir, subject, "mri/aseg.mgz"),
  color_lut = file.path(freesurfer::fs_dir(), "ASegStatsLUT.txt"),
  steps = 1:5,
  output_dir = tempdir(),
  verbose = TRUE,
  ncores = parallel::detectCores() - 2,
  cleanup = TRUE
)
}
\arguments{
\item{subject}{Freesurfer subject, must exist in whatever subject directory
specified or set in the environment with $SUBJECTS_DIR}

\item{subjects_dir}{Freesurfer subject directory}

\item{template}{tempalte volume.mgz file path}

\item{color_lut}{Freesurfer colour look-up-table. Either
as a path or a data.frame that \code{\link{is_ctab}}}

\item{steps}{if cleanup is disabled, all files are
saved, and steps can be re-run individually}

\item{output_dir}{output directory path}

\item{verbose}{logical indicating to be verbose or not}

\item{ncores}{number of cores for parallel processing (default numcores-2)}

\item{cleanup}{logical to toggle removal of all intermediary files}
}
\value{
returns a ggseg3d-atlas ready object
}
\description{
Function to create a ggseg3d-atlas
from a volumetric parcellation.
Currently only tested using \code{.mgz}
extension images. Will call command-line
tools to complete the process.
}
\details{
the function is aggressive in the number
of cores used to minimize time taken to
create the atlas. Set the  \code{ncores}
argument for less aggressive approach.
}
