% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-ggseg-atlas.R
\name{make_volumetric_ggseg}
\alias{make_volumetric_ggseg}
\title{Make ggseg atlas from volumetric template}
\usage{
make_volumetric_ggseg(
  label_file,
  subject = "fsaverage5",
  subjects_dir = fs_subj_dir(),
  output_dir = tempdir(),
  color_lut = NULL,
  steps = 1:8,
  skip_existing = TRUE,
  slices = data.frame(x = c(130, 122, 122), y = c(130, 235, 112), z = c(130, 100, 106),
    view = c("axial", "sagittal", "coronal"), stringsAsFactors = FALSE),
  vertex_size_limits = NULL,
  dilate = NULL,
  tolerance = 0,
  smoothness = 5,
  verbose = TRUE,
  cleanup = FALSE
)
}
\arguments{
\item{label_file}{a volumetric image containing the labels}

\item{subject}{Freesurfer subject, must exist in whatever subject directory
specified or set in the environment with $SUBJECTS_DIR}

\item{subjects_dir}{Freesurfer subject directory}

\item{output_dir}{output directory path}

\item{color_lut}{a file containing color information for the labels}

\item{steps}{numeric vector of steps to run}

\item{skip_existing}{logical. If slice snapshots already exist, should these be skipped.}

\item{slices}{a data.frame with columns x, y, z, and view specifying coordinates and view of slice snapshots.}

\item{vertex_size_limits}{numeric vector of two, setting the minimum and maximum vector size of polygons. Defaults to NULL, which sets no limits.}

\item{dilate}{numeric. Dilation factor for polygons. Default NULL applies no dilation.}

\item{tolerance}{tolerance during vertex reduction \code{\link[sf]{st_simplify}}}

\item{smoothness}{smoothing factor, argument to \code{\link[smoothr]{smooth}}}

\item{verbose}{logical indicating to be verbose or not}

\item{cleanup}{logical to toggle removal of all intermediary files}
}
\value{
brain-atlas class
}
\description{
If making an atlas from a non-cortical atlas,
volumetric atlases are the best options.
Instead of snapshotting images of inflated
brain, will snapshot brain slices given x, y, z
coordinates for the slices trhoush the \code{slices}
argument.
}
\examples{
\dontrun{

   label_file <- file.path(fs_subj_dir(), subject, "mri/aseg.mgz")
   slices = data.frame(x=130, y=130, z=130, view="axial", stringsAsFactors = FALSE)

   aseg2 <- make_volumetric_ggseg(
      label_file =  label_file,
      slices = slices
   )

   # Have a look at the atlas
   plot(aseg2)
}
}
