% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify-atlas.R
\name{match_vertices}
\alias{match_vertices}
\title{Match region vertices to brain mesh vertices}
\usage{
match_vertices(region_coords, brain_coords, tolerance = 1e-04)
}
\description{
Match region vertices to brain mesh vertices
}
\keyword{internal}
