% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{mri_annotation2label}
\alias{mri_annotation2label}
\title{Convert annotation to label}
\usage{
mri_annotation2label(
  annot_name,
  subject = "fsaverage5",
  hemisphere = "rh",
  output_dir = fs_subj_dir(),
  verbose = TRUE,
  opts = NULL
)
}
\arguments{
\item{annot_name}{annotation name. File should exist in subjects label directory}

\item{subject}{Freesurfer subject, must exist in whatever subject directory
specified or set in the environment with $SUBJECTS_DIR}

\item{hemisphere}{hemisphere, either "rh" or "lh"}

\item{output_dir}{output directory path}

\item{verbose}{logical indicating to be verbose or not}

\item{opts}{other arguments to freesurfer command}
}
\value{
nothing. Runs command line to write label files
}
\description{
Calls FreeSurfer's \code{mri_annotation2label}
to split an annotation file into several labels.
}
\examples{
\dontrun{
# for freesurfer help see:
freesurfer::fs_help("mri_annotation2label")
mri_annotation2label(annot_name = "aparc")

mri_annotation2label(annot_name = "aparc.a2009s")

mri_annotation2label(subject = "fsaverage", annot_name = "aparc.a2009s")
}
}
