% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{mri_pretess}
\alias{mri_pretess}
\title{Run pre-tesselation on file}
\usage{
mri_pretess(template, label, output_file, verbose = TRUE, opts = NULL)
}
\arguments{
\item{template}{template mgz}

\item{label}{label to run}

\item{output_file}{path to output file}

\item{verbose}{logical indicating to be verbose or not}

\item{opts}{other arguments to freesurfer command}
}
\description{
Run pre-tesselation on file
}
