% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{mri_tessellate}
\alias{mri_tessellate}
\title{Tesselate data}
\usage{
mri_tessellate(input_file, label, output_file, verbose)
}
\arguments{
\item{input_file}{input file}

\item{label}{label to run}

\item{output_file}{path to output file}

\item{verbose}{logical indicating to be verbose or not}
}
\description{
Tesselate data
}
