% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{mri_vol2surf}
\alias{mri_vol2surf}
\title{Convert volume to surface}
\usage{
mri_vol2surf(
  input_file,
  output_file,
  hemisphere,
  projfrac = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{input_file}{input volume}

\item{output_file}{path to output file}

\item{hemisphere}{hemisphere, either "rh" or "lh"}

\item{projfrac}{argument to projfrac}

\item{verbose}{logical indicating to be verbose or not}
}
\description{
Convert volume to surface
}
