% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfsplit.R
\name{prep_n_data}
\alias{prep_n_data}
\title{prep data for verteces/faces}
\usage{
prep_n_data(n, type = "vertex")
}
\arguments{
\item{n}{unnused arg, need for apply}

\item{type}{vertex or face}
}
\description{
prep data for verteces/faces
}
