% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_annotation.R
\name{read_annotation}
\alias{read_annotation}
\title{Read Freesurfer annotation file}
\usage{
read_annotation(path, verbose = TRUE)
}
\arguments{
\item{path}{path to annotation file}

\item{verbose}{logical.}
}
\value{
dataframe with 6 variables
}
\description{
Reads Freesurfer binary annotation files
that contain information on vertex labels
and colours for use in analyses and
brain area lookups.
}
\details{
This function is heavily
based on Freesurfer's read_annotation.m
Original Author: Bruce Fischl
CVS Revision Info:
    $Author: greve $
    $Date: 2014/02/25 19:54:10 $
    $Revision: 1.10 $

Copyright © 2011 The General Hospital Corporation (Boston, MA) "MGH"
Terms and conditions for use, reproduction, distribution and contribution
are found in the 'FreeSurfer Software License Agreement' contained
in the file 'LICENSE' found in the FreeSurfer distribution, and here:
https://surfer.nmr.mgh.harvard.edu/fswiki/FreeSurferSoftwareLicense
Reporting: freesurfer@nmr.mgh.harvard.edu

 # TODO: replace with freesurfer::read_annotation when its incorporated on CRAN
}
