% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colortable-funcs.R
\name{read_ctab}
\alias{read_ctab}
\title{Read colourtab}
\usage{
read_ctab(path)
}
\arguments{
\item{path}{path to read from}
}
\value{
a data.frame with index, label name and RGBA colours
}
\description{
Read in a FreeSurfer colortab file.
}
