% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_dpv}
\alias{read_dpv}
\title{Read dpv file}
\usage{
read_dpv(path)
}
\arguments{
\item{path}{path to dpv file}
}
\value{
list of vertices and faces
}
\description{
Read in data-per-vertex file, which is
usually a file made by for instance \code{\link{surf2asc}},
\code{\link{curv2asc}}, \code{\link{curvnf2asc}}
}
