% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{read_label_vertices}
\alias{read_label_vertices}
\title{Read vertex indices from a FreeSurfer label file}
\usage{
read_label_vertices(label_file)
}
\arguments{
\item{label_file}{Path to .label file}
}
\value{
Integer vector of vertex indices (0-indexed)
}
\description{
Read vertex indices from a FreeSurfer label file
}
\keyword{internal}
