% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{read_ply_mesh}
\alias{read_ply_mesh}
\title{Read mesh data from PLY file}
\usage{
read_ply_mesh(ply, ...)
}
\arguments{
\item{ply}{path to ply-file}

\item{...}{arguments to \code{\link[geomorph:read.ply]{geomorph::read.ply()}}}
}
\value{
list with vertices (data.frame with x, y, z) and
faces (data.frame with i, j, k)
}
\description{
PLY files contain mesh information including vertices and faces.
This function reads a PLY file and extracts the vertices and faces
into a list format suitable for ggseg3d.
}
\keyword{internal}
