% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{read_volume}
\alias{read_volume}
\title{Read neuroimaging volume file}
\usage{
read_volume(file)
}
\arguments{
\item{file}{Path to volume file (.mgz, .nii, .nii.gz)}
}
\value{
3D array of volume data
}
\description{
Reads volume data from common neuroimaging formats including
FreeSurfer MGZ and NIfTI.
}
\keyword{internal}
