% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-convert.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{convert_legacy_brain_atlas}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggseg.formats}{\code{\link[ggseg.formats]{convert_legacy_brain_atlas}}}
}}

