% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract-geometry.R
\name{resample_streamline}
\alias{resample_streamline}
\title{Resample streamline to fixed number of points}
\usage{
resample_streamline(streamline, n_points)
}
\description{
Resample streamline to fixed number of points
}
\keyword{internal}
