% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-tracts.R
\name{resolve_tube_radius}
\alias{resolve_tube_radius}
\title{Resolve tube radius specification}
\usage{
resolve_tube_radius(tube_radius, streamlines, centerline, density_range)
}
\description{
Resolve tube radius specification
}
\keyword{internal}
