% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restruct_old_3datlas}
\alias{restruct_old_3datlas}
\title{Change old atlas setup to new}
\usage{
restruct_old_3datlas(atlas_data)
}
\arguments{
\item{atlas_data}{ggseg3d-atlas object}
}
\description{
Change old atlas setup to new
}
