% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract-geometry.R
\name{rotate_vector}
\alias{rotate_vector}
\title{Rotate vector around axis by angle (Rodrigues' formula)}
\usage{
rotate_vector(v, axis, angle)
}
\description{
Rotate vector around axis by angle (Rodrigues' formula)
}
\keyword{internal}
