% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaefer.R
\docType{data}
\name{schaefer17_200}
\alias{schaefer17_200}
\title{Local-Global 17 Parcellation of the Human Cerebral Cortex}
\format{A data.frame
\describe{
  \item{long}{coordinates for the x-axis}
  \item{lat}{coordinates for the y-axis}
  \item{area}{name of network}
  \item{hemi}{name of the hemisphere (left, right)}
  \item{side}{which side to view (medial, lateral)}
  \item{network}{network number (1:7)}
  \item{label}{unique name to each node}
  \item{atlas}{name of the atlas}
}}
\usage{
data(schaefer17_200)
}
\description{
Polygon data for the 17 resting-state networks of
the Schaefer et al. (2018) networks, in a 200 area parcellation
}
\examples{
data(schaefer17_200)
}
\references{
Schaefer et al. (2018) Cereb Cortex. 2018 Sep 1;28(9):3095-3114. doi: 10.1093/cercor/bhx179
(\href{https://www.ncbi.nlm.nih.gov/pubmed/28981612}{PubMed})
}
\seealso{
Other ggseg_atlases: \code{\link{chenAr}},
  \code{\link{chenTh}}, \code{\link{glasser}},
  \code{\link{hoCort}}, \code{\link{jhu}},
  \code{\link{tracula}}, \code{\link{yeo17}},
  \code{\link{yeo7}}
}
\concept{ggseg_atlases}
\keyword{datasets}
