% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaefer.R
\docType{data}
\name{schaefer7_3d}
\alias{schaefer7_3d}
\title{Local-Global 7 Parcellation of the Human Cerebral Cortex}
\format{A tibble with 4 observations and a nested data.frame
\describe{
  \item{surf}{type of surface (`inflated` or `white`)}
  \item{hemi}{hemisphere (`left`` or `right`)}
  \item{data}{data.frame of necessary variables for plotting
  }

  \item{atlas}{String. atlas name}
  \item{roi}{numbered region from surface}
  \item{annot}{concatenated region name}
  \item{label}{label `hemi_annot` of the region}
  \item{mesh}{list of meshes in two lists: vb and it}
  \item{area}{name of area in full}
  \item{colour}{HEX colour of region}
}}
\usage{
data(schaefer7_3d)
}
\description{
Mesh data for the 7 resting-state networks of
the Schaefer et al. (2018) networks.
}
\examples{
data(schaefer7_3d)
}
\references{
Schaefer et al. (2018) Cereb Cortex. 2018 Sep 1;28(9):3095-3114. doi: 10.1093/cercor/bhx179
(\href{https://www.ncbi.nlm.nih.gov/pubmed/28981612}{PubMed})
}
\seealso{
Other ggseg3d_atlases: \code{\link{desterieux_3d}},
  \code{\link{glasser_3d}}, \code{\link{hc_pa_3d}},
  \code{\link{icbm_3d}}, \code{\link{jhu_3d}},
  \code{\link{schaefer17_3d}}, \code{\link{tracula_3d}},
  \code{\link{yeo17_3d}}, \code{\link{yeo7_3d}}
}
\concept{ggseg3d_atlases}
\keyword{datasets}
