% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_verbosity}
\alias{set_verbosity}
\title{Set verbosity level}
\usage{
set_verbosity(level)
}
\arguments{
\item{level}{Verbosity level: 0 (silent), 1 (normal), 2 (high)}
}
\value{
Previous verbosity level (invisibly)
}
\description{
Set verbosity level
}
\examples{
# Set to silent
set_verbosity(0)

# Set to normal
set_verbosity(1)

# Set to high (includes FreeSurfer verbose output)
set_verbosity(2)
}
