% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-steps.R
\name{snapshot_brain}
\alias{snapshot_brain}
\title{Make snapshots through orca and plotly}
\usage{
snapshot_brain(ggseg3d_atlas, hemisphere, view, surface, output_dir, pb = NULL)
}
\arguments{
\item{ggseg3d_atlas}{object of class ggseg3d-atlas}

\item{hemisphere}{hemisphere, either "rh" or "lh"}

\item{view}{view}

\item{surface}{Freesurfer surface}

\item{output_dir}{output directory path}

\item{pb}{progressbar}
}
\description{
Make snapshots through orca and plotly
}
