% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject_2_ascii.R
\name{subject_2_ascii}
\alias{subject_2_ascii}
\title{Convert subject surface files to ascii}
\usage{
subject_2_ascii(
  subject = "fsaverage5",
  hemisphere = c("rh", "lh"),
  surfaces = fs_surfaces(),
  curvatures = fs_curvatures(),
  nofix_curv = fs_nofixcurv(),
  subjects_dir = freesurfer::fs_subj_dir(),
  output_dir = subjects_dir,
  verbose = TRUE
)
}
\arguments{
\item{subject}{Freesurfer subject, must exist in whatever subject directory
specified or set in the environment with $SUBJECTS_DIR}

\item{hemisphere}{hemisphere, either "rh" or "lh"}

\item{surfaces}{string vector of surfaces}

\item{curvatures}{string vector of curvatures}

\item{nofix_curv}{string vector of nofix curvatures}

\item{subjects_dir}{Freesurfer subject directory}

\item{output_dir}{output directory path}

\item{verbose}{logical indicating to be verbose or not}
}
\description{
This function goes through all specified
subject,  per hermisphere, surface,
curvature and no fix curvature specified
and turns them into ascii files.
}
