% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{surf2asc}
\alias{surf2asc}
\title{Convert Freesurfer surface file to ascii}
\usage{
surf2asc(input_file, output_file, verbose = TRUE)
}
\arguments{
\item{input_file}{path to input surface file to convert}

\item{output_file}{path to output file}

\item{verbose}{logical indicating to be verbose or not}
}
\value{
ascii data
}
\description{
Surface files from Freesurfer need to be
conserted to be handled in R. This first
step turns the surface file in to a text
ascii file.
}
