% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfsplit.R
\name{surfsplit}
\alias{surfsplit}
\title{Split surface file into separate plys}
\usage{
surfsplit(
  srf_ply,
  label_path,
  prefix = "test",
  output_dir = dirname(label_path),
  verbose = TRUE
)
}
\arguments{
\item{srf_ply}{Surface ply path}

\item{label_path}{label dpv file}

\item{prefix}{output prefix}

\item{output_dir}{output directory path}

\item{verbose}{logical indicating to be verbose or not}
}
\value{
list of ply meshes
}
\description{
Function to splot up a surface ply
into several ply's based on a freesurfer
label file.
}
\details{
Script adapted to R and ggseg based on scripts
from Anderson M. Winkler: http://brainder.org
}
