% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracula.R
\docType{data}
\name{tracula}
\alias{tracula}
\title{Automated probabilistic reconstruction of white-matter pathways}
\format{A data.frame with 11341 observations and 11 variables
\describe{
  \item{long}{coordinates for the x-axis}
  \item{lat}{coordinates for the y-axis}
  \item{area}{name of region}
  \item{hemi}{dummy name of the hemisphere}
  \item{side}{which side to view (sagittal)}
  \item{label}{label name from Freesurfer}
  \item{atlas}{name of the atlas}
}}
\usage{
tracula
}
\description{
Automated probabilistic reconstruction of white-matter pathways
}
\examples{
data(tracula)
}
\references{
Yendiki et al. (2011) Automated probabilistic reconstruction of 
white-matter pathways in health and disease using an atlas of the underlying anatomy.  
Front. Neuroinform. 5:23. doi: 10.3389/fninf.2011.00023
(\href{https://www.ncbi.nlm.nih.gov/pubmed/22016733}{PubMed})
}
\seealso{
Other ggseg_atlases: 
\code{\link{chenAr}},
\code{\link{chenTh}},
\code{\link{dkextra}},
\code{\link{glasser}},
\code{\link{hcpa_3d}},
\code{\link{hoCort}},
\code{\link{jhu}},
\code{\link{yeo17}},
\code{\link{yeo7}}
}
\concept{ggseg_atlases}
\keyword{datasets}
