% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracula.R
\docType{data}
\name{tracula_3d}
\alias{tracula_3d}
\title{White matter tract parcellations}
\format{An object of class \code{ggseg_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.}
\usage{
tracula_3d
}
\description{
White matter tract parcellations
}
\examples{
data(tracula_3d)
}
\references{
Yendiki et al. (2011) Automated probabilistic reconstruction of 
white-matter pathways in health and disease using an atlas of the underlying anatomy.  
Front. Neuroinform. 5:23. doi: 10.3389/fninf.2011.00023
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2724595/}{PubMed})
}
\keyword{datasets}
