% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash-exec.R
\name{wb_command}
\alias{wb_command}
\title{HPC workbench command}
\usage{
wb_command(wbdir = NULL)
}
\arguments{
\item{wbdir}{optional path to workbench program}
}
\value{
PATH
}
\description{
Some options for optimization of region
extraction for ggseg atlases can be run using the
HPC workbench program. This can be installed on the system
in various way. Here, the function will either assume
workbench is in standard macOS location, or in the locatino
you provide to the function call. This path will be added
to the environment $PATH. Currently does not work on windows.
}
\examples{
wb_command()

wb_command("some/path/workbench")
}
