% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{write_ctab}
\alias{write_ctab}
\title{Write FreeSurfer color table}
\usage{
write_ctab(x, path)
}
\arguments{
\item{x}{A data.frame with columns: idx, label, R, G, B, A.}

\item{path}{Path to write to.}
}
\value{
Invisibly returns the lines written.
}
\description{
Write a color table to file in FreeSurfer format.
}
\seealso{
\code{\link[=read_ctab]{read_ctab()}}, \code{\link[=is_ctab]{is_ctab()}}
}
