% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_dpv}
\alias{write_dpv}
\title{Write DPV file}
\usage{
write_dpv(path, vertices, faces)
}
\arguments{
\item{path}{path to file}

\item{vertices}{object with vertices}

\item{faces}{object with faces}
}
\description{
write a data-per-vertex file,
and ascii-format text-file
based on vertices and faces.
}
