% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-steps.R
\name{atlas_lab2ctab}
\alias{atlas_lab2ctab}
\title{Label to ctab}
\usage{
atlas_lab2ctab(output_dir, verbose)
}
\arguments{
\item{output_dir}{output directory path}

\item{verbose}{logical indicating to be verbose or not}
}
\description{
Label to ctab
}
