% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-steps.R
\name{atlas_vol2label}
\alias{atlas_vol2label}
\title{Volume to label}
\usage{
atlas_vol2label(annot_lab, output_dir, verbose)
}
\arguments{
\item{annot_lab}{annotation label}

\item{output_dir}{output directory path}

\item{verbose}{logical indicating to be verbose or not}
}
\description{
Volume to label
}
