% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas-steps.R
\name{atlas_vol2surf}
\alias{atlas_vol2surf}
\title{Nifti volume to surface}
\usage{
atlas_vol2surf(input_file, output_dir, projfrac = 0.5, verbose = TRUE)
}
\arguments{
\item{input_file}{nifti volume}

\item{output_dir}{output directory path}

\item{projfrac}{value to mri_vol2surf -projfrac}

\item{verbose}{logical indicating to be verbose or not}
}
\description{
Nifti volume to surface
}
