% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{curv2ply}
\alias{curv2ply}
\title{Convert Freesurfer curvature file to ply}
\usage{
curv2ply(input_file, output_file = paste(input_file, ".ply"), verbose = TRUE)
}
\arguments{
\item{input_file}{path to Freesurfer curvature file}

\item{output_file}{path to output file}

\item{verbose}{logical indicating to be verbose or not}
}
\value{
.ply text
}
\description{
Function to convert Freesurfer curvature
file into .ply
}
