% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-structs.R
\name{fs_nofixcurvatures}
\alias{fs_nofixcurvatures}
\alias{fs_nofixcurv}
\title{Available Freesurfer no fix curvatures}
\usage{
fs_nofixcurvatures()

fs_nofixcurv()
}
\value{
character vector of no-fix curvature file names
}
\description{
Available Freesurfer no fix curvatures
}
\section{Functions}{
\itemize{
\item \code{fs_nofixcurv()}: Deprecated. Use \code{\link[=fs_nofixcurvatures]{fs_nofixcurvatures()}} instead.

}}
\examples{
fs_nofixcurvatures()
}
