% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colortable-funcs.R
\name{get_ctab}
\alias{get_ctab}
\title{Get colour lut table}
\usage{
get_ctab(color_lut)
}
\arguments{
\item{color_lut}{path to lut or data.frame that \code{\link{is_ctab}}}
}
\value{
colour table
}
\description{
Get colour lut table
}
