% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-atlases.R
\name{ggseg_atlas_repos}
\alias{ggseg_atlas_repos}
\title{List all online repositories with ggseg-atlases}
\usage{
ggseg_atlas_repos(pattern = NULL, ...)
}
\arguments{
\item{pattern}{string pattern to search repos}

\item{...}{additional arguments to \code{\link[base]{grep}}}
}
\value{
data frame of online repositories with ggseg-atlases
}
\description{
Function to easily find all online repositories that
contain ggseg atlases
}
\examples{
ggseg_atlas_repos()

ggseg_atlas_repos("yeo")
}
