% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desterieux.R
\docType{data}
\name{hc_pa_3d}
\alias{hc_pa_3d}
\title{Hippocampal posterior-anterior segmentation}
\format{A tibble with 4 observations and a nested data.frame
\describe{
  \item{surf}{type of surface (`inflated` or `white`)}
  \item{hemi}{hemisphere (`left`` or `right`)}
  \item{data}{data.frame of necessary variables for plotting
  }

  \item{atlas}{String. atlas name}
  \item{roi}{numbered region from surface}
  \item{annot}{concatenated region name}
  \item{label}{label `hemi_annot` of the region}
  \item{mesh}{list of meshes in two lists: vb and it}
  \item{area}{name of area in full}
  \item{colour}{HEX colour of region}
}}
\usage{
data(hc_pa_3d)
}
\description{
Mesh data for the anterior and posterior 
segmentations of the hippocampus
}
\examples{
data(hc_pa_3d)
}
\seealso{
Other ggseg3d_atlases: \code{\link{desterieux_3d}},
  \code{\link{glasser_3d}}, \code{\link{icbm_3d}},
  \code{\link{jhu_3d}}, \code{\link{schaefer17_3d}},
  \code{\link{schaefer7_3d}}, \code{\link{tracula_3d}},
  \code{\link{yeo17_3d}}, \code{\link{yeo7_3d}}
}
\concept{ggseg3d_atlases}
\keyword{datasets}
