% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icbm.R
\docType{data}
\name{icbm_3d}
\alias{icbm_3d}
\title{Parcellation from of white matter}
\format{An object of class \code{ggseg_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.}
\usage{
icbm_3d
}
\description{
Parcellation from of white matter
}
\examples{
data(icbm_3d)
}
\references{
Mori et al., MRI Atlas of Human White Matter. Elsevier, Amsterdam, The Netherlands (2005) 
(\href{https://www.elsevier.com/books/mri-atlas-of-human-white-matter/mori/978-0-444-51741-8}{PubMed})
}
\keyword{datasets}
