% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-atlases.R
\name{install_ggseg_atlas}
\alias{install_ggseg_atlas}
\title{Install a ggseg atlas package}
\usage{
install_ggseg_atlas(
  package,
  repos = c(ggsegverse = "https://ggsegverse.r-universe.dev", getOption("repos")),
  ...
)
}
\arguments{
\item{package}{Package name (e.g., \code{"ggsegYeo2011"}, \code{"ggsegDKT"}).
Use \code{\link[ggseg.hub:ggseg_atlas_repos]{ggseg_atlas_repos()}} to see available packages.}

\item{repos}{Repositories to install from. Defaults to the ggseg
r-universe plus your current CRAN mirror.}

\item{...}{Additional arguments passed to \code{\link[utils:install.packages]{utils::install.packages()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has moved to the ggseg.hub package. Please use
\code{\link[ggseg.hub:install_ggseg_atlas]{ggseg.hub::install_ggseg_atlas()}} instead.
}
