% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-atlases.R
\name{install_ggseg_atlas}
\alias{install_ggseg_atlas}
\title{Install ggseg-atlas from repo}
\usage{
install_ggseg_atlas(
  package,
  repos = c(ggseg = "https://ggseg.r-universe.dev", CRAN =
    "https://cloud.r-project.org"),
  ...
)
}
\arguments{
\item{package}{package name}

\item{repos}{vector of repositories to install from. Defaults to
ggseg r-universe and CRAN.}

\item{...}{additional arguments to \code{install.packages}.}
}
\description{
installs ggseg-atlas library from
the ggseg r-universe <https://ggseg.r-universe.dev/ui#builds> .
}
\details{
To install, will temporarily alter your install
repo settings to also use the ggseg r-universe
build as source for packages. These settings will
be restored when the function exits.
}
\examples{
\dontrun{
yeo2011_repo <- ggseg_atlas_repos("yeo2011")
install_ggseg_atlas(yeo2011_repo$repo, yeo2011_repo$source)
}
}
