% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-atlases.R
\name{install_ggseg_atlas}
\alias{install_ggseg_atlas}
\title{Install ggseg-atlas from repo}
\usage{
install_ggseg_atlas(repo, source, ...)
}
\arguments{
\item{repo}{repository string}

\item{source}{source of repository as string}

\item{...}{additional arguments to \code{remotes::install_[source]}
function called depending on the \code{source}.}
}
\description{
installs ggseg-atlas library from
online repository.
}
\examples{
\dontrun{
yeo2011_repo <- ggseg_atlas_repos("yeo2011")
install_ggseg_atlas(yeo2011_repo$repo, yeo2011_repo$source)
}
}
