% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-atlases.R
\name{install_ggseg_atlas_all}
\alias{install_ggseg_atlas_all}
\title{Install all available ggseg atlas packages}
\usage{
install_ggseg_atlas_all(
  repos = c(ggseg = "https://ggseg.r-universe.dev", getOption("repos")),
  ...
)
}
\arguments{
\item{repos}{Repositories to install from. Defaults to the ggseg
r-universe plus your current CRAN mirror.}

\item{...}{Additional arguments passed to \code{\link[utils:install.packages]{utils::install.packages()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has moved to the ggseg.hub package. Please use
\code{\link[ggseg.hub:install_ggseg_atlas_all]{ggseg.hub::install_ggseg_atlas_all()}} instead.
}
