% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-atlases.R
\name{install_ggseg_atlas_all}
\alias{install_ggseg_atlas_all}
\title{Install all ggseg-atlases registeres}
\usage{
install_ggseg_atlas_all(...)
}
\arguments{
\item{...}{additional arguments to \code{remotes::install_*}
function called depending on the \code{source}.}
}
\description{
Calls \code{\link{ggseg_atlas_repos}}, 
and installs all libraries listed.
Be careful calling this function, this
will likely take quite some time to
complete, and also will contain a lot of
heavy data. We recommend only installing 
atlases you are likely to use and on demand.
}
\examples{
\dontrun{
install_ggseg_atlas_all()
}
}
