% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colortable-funcs.R
\name{is_ctab}
\alias{is_ctab}
\title{Check if object is colourtable}
\usage{
is_ctab(colourtable)
}
\arguments{
\item{colourtable}{data frame with colour table}
}
\value{
logical
}
\description{
Check if object is colourtable
}
