% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freesurfer.R
\name{lcbc_surf2surf}
\alias{lcbc_surf2surf}
\title{Convert LCBC surface file to other subjects}
\usage{
lcbc_surf2surf(
  input_volume,
  source_subject = "fsaverage",
  target_subject = "fsaverage5",
  hemisphere = "rh",
  subjects_dir = fs_subj_dir(),
  output_dir = file.path(subjects_dir, target_subject, "surf"),
  cortex = TRUE,
  verbose = get_verbose()
)
}
\arguments{
\item{input_volume}{path to input volume}

\item{source_subject}{source subject}

\item{target_subject}{target subject}

\item{hemisphere}{hemisphere, either "rh" or "lh"}

\item{subjects_dir}{Freesurfer subject directory}

\item{output_dir}{Directory to store intermediate files (screenshots, masks,
contours). Defaults to \code{\link[=tempdir]{tempdir()}}.}

\item{cortex}{toggle "--cortex" (TRUE) or "--no-cortex" (FALSE)}

\item{verbose}{Logical. If not specified, uses the value from
\code{options("ggsegExtra.verbose")} or the \code{GGSEGEXTRA_VERBOSE}
environment variable. Defaults to \code{TRUE}.
Use \code{options(ggsegExtra.verbose = FALSE)} to suppress output globally.}
}
\description{
Convert LCBC surface file to other subjects
}
\keyword{internal}
