% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{lcbc_surf2surf}
\alias{lcbc_surf2surf}
\title{Convert LCBC surface file to other subjects}
\usage{
lcbc_surf2surf(
  input_volume,
  source_subject = "fsaverage",
  target_subject = "fsaverage5",
  hemisphere = "rh",
  subjects_dir = freesurfer::fs_subj_dir(),
  output_dir = file.path(subjects_dir, target_subject, "surf"),
  cortex = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{input_volume}{path to input volume}

\item{source_subject}{source subject}

\item{target_subject}{target subject}

\item{hemisphere}{hemisphere, either "rh" or "lh"}

\item{subjects_dir}{Freesurfer subject directory}

\item{output_dir}{output directory path}

\item{cortex}{toggle "--cortex" (TRUE) or "--no-cortex" (FALSE)}

\item{verbose}{logical indicating to be verbose or not}
}
\description{
Convert LCBC surface file to other subjects
}
