% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-ggseg-atlas.R
\name{make_ggseg3d_2_ggseg}
\alias{make_ggseg3d_2_ggseg}
\title{Turn ggseg3d-atlas to ggseg}
\usage{
make_ggseg3d_2_ggseg(
  ggseg3d_atlas = ggseg3d::dk_3d,
  steps = 1:6,
  output_dir = tempdir(),
  smoothness = 10,
  keep = 0.05,
  weighting = 0.7,
  cleanup = FALSE
)
}
\arguments{
\item{ggseg3d_atlas}{object of class ggseg3d-atlas}

\item{steps}{number of 1:6 of which steps to run through}

\item{output_dir}{output directory path}

\item{smoothness}{smoothing factor, argument to \code{\link[smoothr]{smooth}}}

\item{keep}{percent of vertices to keep during vertex reduction \code{\link[rmapshaper]{ms_simplify}}}

\item{weighting}{weighting of vertices for vertex reduction \code{\link[rmapshaper]{ms_simplify}}}

\item{cleanup}{logical to toggle removal of all intermediary files}
}
\value{
data.frame ready for manual cleaning
}
\description{
Function will create a dataframe
based on a ggseg3d atlas, based on
the contours of each segment.
}
