% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-ggseg3d-atlas.R
\name{make_volumetric_2_3datlas}
\alias{make_volumetric_2_3datlas}
\title{Volumetric segmentation to 3d-atlas}
\usage{
make_volumetric_2_3datlas(
  template,
  color_lut,
  subject = "fsaverage5",
  subjects_dir = freesurfer::fs_subj_dir(),
  steps = 1:5,
  output_dir = tempdir(),
  verbose = TRUE,
  ncores = parallel::detectCores() - 2,
  cleanup = TRUE
)
}
\arguments{
\item{template}{template \code{volume.mgz} file path}

\item{color_lut}{Freesurfer colour look-up-table. Either
as a path or a data.frame that \code{\link{is_ctab}}}

\item{subject}{Freesurfer subject, must exist in whatever subject directory
specified or set in the environment with $SUBJECTS_DIR}

\item{subjects_dir}{Freesurfer subject directory}

\item{steps}{if cleanup is disabled, all files are
saved, and steps can be re-run individually}

\item{output_dir}{output directory path}

\item{verbose}{logical indicating to be verbose or not}

\item{ncores}{number of cores for parallel processing (default numcores-2)}

\item{cleanup}{logical to toggle removal of all intermediary files}
}
\value{
returns a ggseg3d-atlas ready object. Might need manual
   cleaning to become a good atlas.
}
\description{
Function to create a ggseg3d-atlas
from a volumetric parcellation.
Currently only tested using \code{.mgz}
extension images. Will call command-line
tools to complete the process.
}
\details{
the function is aggressive in the number
of cores used to minimize time taken to
create the atlas. Set the  \code{ncores}
argument for less aggressive approach.
}
\examples{
\dontrun{

fs_subject_dir <- freesurfer::fs_dir()
aseg_temp <- file.path(fs_subject_dir, "fsaverage5/mri/aseg.mgz")
colorlut <- file.path(fs_subject_dir, "ASegStatsLUT.txt")

make_volumetric_2_3datlas(aseg_temp, colorlut)
}
}
