% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{mri_vol2label}
\alias{mri_vol2label}
\title{Convert volume to label}
\usage{
mri_vol2label(
  input_file,
  label_id,
  hemisphere,
  subject = "fsaverage5",
  subjects_dir = freesurfer::fs_subj_dir(),
  output_dir,
  verbose = TRUE
)
}
\arguments{
\item{input_file}{input volume}

\item{label_id}{label to run}

\item{hemisphere}{hemisphere, either "rh" or "lh"}

\item{subject}{Freesurfer subject, must exist in whatever subject directory
specified or set in the environment with $SUBJECTS_DIR}

\item{subjects_dir}{Freesurfer subject directory}

\item{output_dir}{output directory path}

\item{verbose}{logical indicating to be verbose or not}
}
\description{
Convert volume to label
}
