% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{mris_label2annot}
\alias{mris_label2annot}
\title{Convert Label to Annotation}
\usage{
mris_label2annot(
  labels,
  hemisphere = "rh",
  ctab,
  subject = "fsaverage5",
  subjects_dir = freesurfer::fs_subj_dir(),
  annot_dir = file.path(subjects_dir, subject, "label"),
  output_dir = subjects_dir,
  opts = NULL,
  verbose = TRUE
)
}
\arguments{
\item{labels}{label vector}

\item{hemisphere}{hemisphere, either "rh" or "lh"}

\item{ctab}{colourtable file}

\item{subject}{Freesurfer subject, must exist in whatever subject directory
specified or set in the environment with $SUBJECTS_DIR}

\item{subjects_dir}{Freesurfer subject directory}

\item{annot_dir}{path to directory with annot-files}

\item{output_dir}{output directory path}

\item{opts}{other arguments to freesurfer command}

\item{verbose}{logical indicating to be verbose or not}
}
\description{
Convert Label to Annotation
}
