% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_2_atlas.R
\name{nifti_2_atlas}
\alias{nifti_2_atlas}
\title{Convert nifti file with labels to atlas}
\usage{
nifti_2_atlas(
  nifti,
  outdir,
  annot_lab,
  atlas_name,
  dilation = 2,
  eroding = 2,
  smoothing = 2,
  steps = 1:5,
  verbose = TRUE
)
}
\arguments{
\item{nifti}{nifti file}

\item{outdir}{output directory (will be created)}

\item{annot_lab}{annotation dataframe}

\item{atlas_name}{name to give atlas}

\item{dilation}{dilation factor}

\item{eroding}{erosion factor}

\item{smoothing}{smoothing factor}

\item{steps}{which steps to run 1:5}

\item{verbose}{verbose or not}
}
\value{
ggseg-atlas
}
\description{
Convert nifti file with labels to atlas
}
\details{
The steps in this function are:
\itemize{
 \item{step 1 - }{\code{\link{atlas_vol2surf}} -  converts volume to surface}
 \item{step 2 - }{\code{\link{atlas_vol2label}} - reads labels from volume}
 \item{step 3 - }{\code{\link{atlas_lab2ctab}} - creates ctab from labels}
 \item{step 4 - }{\code{\link{atlas_tcl}} - runs tkSurfer tcl script}
 \item{step 5 - }{\code{\link{atlas_isolate}} - isolates coloured regions from images}
}
}
