% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colortable-funcs.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read_ctab}
\alias{write_ctab}
\alias{is_ctab}
\alias{get_ctab}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggseg.formats}{\code{\link[ggseg.formats]{get_ctab}}, \code{\link[ggseg.formats]{is_ctab}}, \code{\link[ggseg.formats]{read_ctab}}, \code{\link[ggseg.formats]{write_ctab}}}
}}

