% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{smooth2srf}
\alias{smooth2srf}
\title{Turn smooth file to ascii}
\usage{
smooth2srf(input_file, output_file, verbose)
}
\arguments{
\item{input_file}{input file path}

\item{output_file}{path to output file}

\item{verbose}{logical indicating to be verbose or not}
}
\description{
Turn smooth file to ascii
}
