% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{surf2ply}
\alias{surf2ply}
\title{Convert Freesurfer surface file into ply}
\usage{
surf2ply(input_file, output_file = paste(input_file, ".ply"), verbose = TRUE)
}
\arguments{
\item{input_file}{path to Freesurfer surface file}

\item{output_file}{path to output file}

\item{verbose}{logical indicating to be verbose or not}
}
\value{
ply text
}
\description{
Function to convert Freesurfer surface
file into .ply
}
