% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracula.R
\docType{data}
\name{tracula}
\alias{tracula}
\title{White matter tract parcellations}
\format{A data.frame with 11341 observations and 11 variables
\describe{
  \item{long}{coordinates for the x-axis}
  \item{lat}{coordinates for the y-axis}
  \item{area}{name of region}
  \item{hemi}{dummy name of the hemisphere}
  \item{side}{which side to view (sagittal)}
  \item{label}{label name from Freesurfer}
  \item{atlas}{name of the atlas}
}}
\usage{
tracula
}
\description{
White matter tract parcellations
}
\examples{
data(tracula)
}
\references{
Yendiki et al. (2011) Automated probabilistic reconstruction of 
white-matter pathways in health and disease using an atlas of the underlying anatomy.  
Front. Neuroinform. 5:23. doi: 10.3389/fninf.2011.00023
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2724595/}{PubMed})
}
\keyword{datasets}
