% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshots.R
\name{volume_projection}
\alias{volume_projection}
\title{Create maximum intensity projection of volume}
\usage{
volume_projection(vol, view)
}
\arguments{
\item{vol}{3D array}

\item{view}{"axial", "sagittal", or "coronal"}
}
\value{
2D matrix (projection)
}
\description{
Projects a 3D volume onto a 2D plane by taking the maximum value along
each ray. This shows the full extent of structures from a given view.
}
\keyword{internal}
